Extract buildinfo and catalog into gradle convention
